BOOKKEEPER-224: Fix findbugs in bookkeeper-server component [missing changes] (ivank)
git-svn-id: https://svn.apache.org/repos/asf/zookeeper/bookkeeper/trunk@1339607 13f79535-47bb-0310-9956-ffa450edef68

+++ bookkeeper-server/src/main/java/org/apache/bookkeeper/proto/NIOServerFactory.java
@@ -485,9 +485,8 @@ public class NIOServerFactory extends Thread {
         }
 
         private class CnxnStats {
-            int packetsSent = 0;
-
-            int packetsReceived = 0;
+            long packetsSent = 0;
+            long packetsReceived = 0;
 
             /**
              * The number of requests that have been submitted but not yet
